+++ examples/tickets-addon/frontend/CHANGELOG.md
@@ -1,11 +1,17 @@
 # @metacore-addons/tickets
 
+## 14.0.0
+
+### Patch Changes
+
+- Updated dependencies [e74975a]
+  - @asteby/metacore-sdk@3.6.0
+
 ## 13.0.0
 
 ### Patch Changes
 
-- Updated dependencies
-- Updated dependencies
+- Updated dependencies [5a932ed]
   - @asteby/metacore-sdk@3.5.0
 
 ## 12.0.0