+++ packages/tools/CHANGELOG.md
@@ -1,11 +1,17 @@
 # @asteby/metacore-tools
 
+## 11.0.0
+
+### Patch Changes
+
+- Updated dependencies [e74975a]
+  - @asteby/metacore-sdk@3.6.0
+
 ## 10.0.0
 
 ### Patch Changes
 
-- Updated dependencies
-- Updated dependencies
+- Updated dependencies [5a932ed]
   - @asteby/metacore-sdk@3.5.0
 
 ## 9.0.0